Module kilt_asset_dids::asset::v1

Structs

EvmSmartContractFungibleReference

An asset reference that is identifiable only by an EVM smart contract (e.g., a fungible token). It is a modification of the CAIP-21 spec according to the rules defined in the Asset DID method specification.

EvmSmartContractNonFungibleIdentifier

An asset identifier for an EVM smart contract collection (e.g., an NFT instance). Since the identifier can be up to 78 characters long of an unknown alphabet, this type simply contains the UTF-8 encoding of such an identifier without applying any special parsing/decoding logic.

EvmSmartContractNonFungibleReference

An asset reference that is identifiable by an EVM smart contract and an optional identifier (e.g., an NFT collection or instance thereof). It is a modification of the CAIP-22 spec and CAIP-29 spec according to the rules defined in the Asset DID method specification.

GenericAssetId

A generic asset ID compliant with the CAIP-19 spec that cannot be boxed in any of the supported variants.

GenericAssetIdentifier

A generic asset identifier as defined in the CAIP-19 spec.

GenericAssetNamespace

A generic asset namespace as defined in the CAIP-19 spec. It stores the provided UTF8-encoded namespace without trying to apply any parsing/decoding logic.

GenericAssetReference

A generic asset reference as defined in the CAIP-19 spec.

Slip44Reference

A Slip44 asset reference. It is a modification of the CAIP-20 spec according to the rules defined in the Asset DID method specification.

Enums

AssetId

The Asset ID component as specified in the Asset DID specification.

Constants

ERC20_NAMESPACE

Namespace for Erc20 assets.

ERC721_NAMESPACE

Namespace for Erc721 assets.

ERC1155_NAMESPACE

Namespace for Erc1155 assets.

MAXIMUM_ASSET_IDENTIFIER_LENGTH

The maximum length of a valid asset ID reference.

MAXIMUM_ASSET_ID_LENGTH

The maximum length, including separator symbols, an asset ID can have according to the minimum values defined by the CAIP-19 definition.

MAXIMUM_ASSET_REFERENCE_LENGTH

The maximum length of a valid asset ID reference.

MAXIMUM_NAMESPACE_LENGTH

The maximum length of a valid asset ID namespace.

MINIMUM_ASSET_IDENTIFIER_LENGTH

The minimum length of a valid asset ID identifier.

MINIMUM_ASSET_ID_LENGTH

The minimum length, including separator symbols, an asset ID can have according to the minimum values defined by the CAIP-19 definition.

MINIMUM_ASSET_NAMESPACE_LENGTH

The minimum length of a valid asset ID namespace.

MINIMUM_ASSET_REFERENCE_LENGTH

The minimum length of a valid asset ID reference.

SLIP44_NAMESPACE

Namespace for Slip44 assets.